+++ layers/CMSSchema/packages/custompostmedia-mutations/src/MutationResolvers/PayloadableRemoveFeaturedImageFromCustomPostMutationResolver.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace PoPCMSSchema\CustomPostMediaMutations\MutationResolvers;
 
+use PoPCMSSchema\Media\TypeAPIs\MediaTypeAPIInterface;
 use PoPSchema\SchemaCommons\MutationResolvers\PayloadableMutationResolverTrait;
 use PoP\ComponentModel\Feedback\ObjectTypeFieldResolutionFeedbackStore;
 use PoP\ComponentModel\QueryResolution\FieldDataAccessorInterface;
@@ -13,6 +14,22 @@ class PayloadableRemoveFeaturedImageFromCustomPostMutationResolver extends Remov
 {
     use PayloadableMutationResolverTrait;
     use PayloadableSetOrRemoveFeaturedImageOnCustomPostMutationResolverTrait;
+
+    private ?MediaTypeAPIInterface $mediaTypeAPI = null;
+
+    final public function setMediaTypeAPI(MediaTypeAPIInterface $mediaTypeAPI): void
+    {
+        $this->mediaTypeAPI = $mediaTypeAPI;
+    }
+    final protected function getMediaTypeAPI(): MediaTypeAPIInterface
+    {
+        if ($this->mediaTypeAPI === null) {
+            /** @var MediaTypeAPIInterface */
+            $mediaTypeAPI = $this->instanceManager->getInstance(MediaTypeAPIInterface::class);
+            $this->mediaTypeAPI = $mediaTypeAPI;
+        }
+        return $this->mediaTypeAPI;
+    }
 
     /**
      * Validate the app-level errors when executing the mutation,