+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Added mutations for categories (#2764)
 - Added mutations for tags (#2765)
 - Validate `assign_terms` capability on `setCategory` and `setTag` mutations (#2772)
+- Create a media item using the attachment from an existing media item (#2787)
 - Added predefined persisted queries: (#2774)
   - [PRO] Create missing translation categories for Polylang
   - [PRO] Translate categories for Polylang